+++ Sources/ReaderViewController.h
@@ -42,6 +42,7 @@
 @property (nonatomic, weak, readwrite) id <ReaderViewControllerDelegate> delegate;
 @property (nonatomic) BOOL thumbnailsOnly;
 @property (nonatomic) BOOL displayShadow;
+@property (nonatomic) CGFloat statusBarHeight;
 @property (nonatomic) UIColor *documentBackgroundColor;
 @property (nonatomic) UIColor *thumbnailsBackgroundColor;